refactor: Move types to internal/
