+++ cmd/commands/new.go
@@ -17,8 +17,8 @@ import (
 	"github.com/caffeine-addictt/waku/cmd/options"
 	"github.com/caffeine-addictt/waku/cmd/template"
 	"github.com/caffeine-addictt/waku/cmd/utils"
-	"github.com/caffeine-addictt/waku/cmd/utils/types"
 	"github.com/caffeine-addictt/waku/internal/log"
+	"github.com/caffeine-addictt/waku/internal/types"
 	"github.com/caffeine-addictt/waku/pkg/config"
 	"github.com/charmbracelet/huh"
 	"github.com/spf13/cobra"
