+++ cmd/license/license_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"github.com/caffeine-addictt/waku/cmd/license"
-	"github.com/caffeine-addictt/waku/cmd/utils"
+	"github.com/caffeine-addictt/waku/internal/utils"
 	"github.com/goccy/go-json"
 	"github.com/stretchr/testify/assert"
 )
