+++ cmd/options/new.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/caffeine-addictt/waku/cmd/global"
 	"github.com/caffeine-addictt/waku/cmd/utils"
-	"github.com/caffeine-addictt/waku/cmd/utils/types"
 	"github.com/caffeine-addictt/waku/internal/log"
+	"github.com/caffeine-addictt/waku/internal/types"
 )
 
 const defaultRepo = "https://github.com/caffeine-addictt/waku.git"
