+++ cmd/template/ignore_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/caffeine-addictt/waku/cmd/template"
-	"github.com/caffeine-addictt/waku/cmd/utils/types"
+	"github.com/caffeine-addictt/waku/internal/types"
 	"github.com/stretchr/testify/assert"
 )
