+++ internal/types/value_guard.go
