+++ internal/types/value_guard_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/caffeine-addictt/waku/cmd/utils/types"
+	"github.com/caffeine-addictt/waku/internal/types"
 	"github.com/goccy/go-json"
 	"github.com/stretchr/testify/assert"
 )
