+++ pkg/config/labels.go
@@ -1,6 +1,6 @@
 package config
 
-import "github.com/caffeine-addictt/waku/cmd/utils/types"
+import "github.com/caffeine-addictt/waku/internal/types"
 
 type TemplateLabel []struct {
 	Name  types.CleanString `json:"name"`
