+++ pkg/config/template.go
@@ -3,7 +3,7 @@ package config
 import (
 	"fmt"
 
-	"github.com/caffeine-addictt/waku/cmd/utils/types"
+	"github.com/caffeine-addictt/waku/internal/types"
 )
 
 // The template.json file
